+++ examples/redis-todo/db/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const mongoose = require('mongoose');
+
+mongoose.connect('mongodb://localhost/redis-todo',
+  { useNewUrlParser: true, useCreateIndex: true }
+);